agent/requirements.txt
uagents==0.12.0
requests==2.31.0
python-dotenv==1.0.0
